Config: load erb files.
On Heroku you only have the production environment. If you'd like to use
staging it will always take from the production section in your yaml
file.

Using an erb file we make it possible to add some logic to that file and
thus enable us to have different hosts on production/staging.